Use a different corner for xbanish
The se and sw corners seem to place the cursor over the host system when running
virt-manager, so my s-n keys switch to a workspace on the host instead of the VM.

The ne and nw corners don't seem to have this problem.

The nw corner is a bit annoying because the cursor is over a browser tab, which
shows a pop-up even when the cursor is hidden.